One/requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
python-dotenv
Pillow
django-password-validators